[Leaf] ArcPAK.cs: Move Leaf LZSS compression logic to a separate class (LeafCompression.cs)
